# docs/swagger.yaml
basePath: /api/v1
definitions:
  constant.GenderValue:
    enum:
    - MALE
    - FEMALE
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
  constant.InterestsType:
    enum:
    - relation_ship
    - personality
    - food
    - game
    - sport
    - pet
    - book_movie_music
    type: string
    x-enum-varnames:
    - RelationShipInterestsType
    - PersonalityInterestsType
    - FoodInterestsType
    - GameInterestsType
    - SportInterestsType
    - PetInterestsType
    - BookMovieMusicInterestsType
...
  schemas.Country:
    properties:
      created_at:
        type: string
      id:
        type: integer
      iso_code:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  schemas.CountryDialCode:
    properties:
      country_id:
        type: integer
      created_at:
        type: string
      dial_code:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
  schemas.GetMe:
    properties:
      country:
        $ref: '#/definitions/schemas.Country'
      country_dial_code:
        $ref: '#/definitions/schemas.CountryDialCode'
      created_at:
        type: string
      deleted_at:
        type: string
      easemob_username:
        type: string
      easemob_uuid:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      user_profile:
        $ref: '#/definitions/schemas.GetProfileRes'
    type: object
  schemas.GetProfileRes:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      count_comment:
        type: integer
      easemob_username:
        type: string
      easemob_uuid:
        type: string
      follower_count:
        type: integer
      following_count:
        type: integer
      gender:
        $ref: '#/definitions/constant.GenderValue'
      id:
        type: string
      interests:
        items:
          $ref: '#/definitions/schemas.GetUserInterestsRes'
        type: array
      is_following:
        type: boolean
      like_count:
        type: integer
      nickname:
        type: string
      self_intro:
        type: string
      user_id:
        type: string
      voice:
        type: string
    type: object
  schemas.GetUserInterestsRes:
    properties:
      name:
        type: string
      type:
        $ref: '#/definitions/constant.InterestsType'
    type: object
...
  schemas.PairTokensResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  schemas.RefreshTokenPayload:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  schemas.RegisterPayload:
    properties:
      dial_code:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  schemas.RegisterResponse:
    properties:
      country:
        $ref: '#/definitions/schemas.Country'
      country_dial_code:
        $ref: '#/definitions/schemas.CountryDialCode'
      created_at:
        type: string
      easemob_username:
        type: string
      easemob_uuid:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  schemas.Response:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
  schemas.SignInPayload:
    properties:
      dial_code:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for boilerplate application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: boilerplate API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get user profile by user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetMe'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: Get user profile
      tags:
      - auth
  /auth/refresh_token:
    post:
      consumes:
      - application/json
      description: Refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.RefreshTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PairTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with phone number and password. Also registers
        the user with Easemob chat service.
      parameters:
      - description: Register credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.RegisterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: Register new user
      tags:
      - auth
  /auth/sign_in:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.SignInPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PairTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: Sign in user
      tags:
      - auth
...
securityDefinitions:
  Bearer:
    description: RoomType "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"