# internal/schemas/auth.go
package schemas

import (
	"time"

	"github.com/google/uuid"
)

type RegisterPayload struct {
	DialCode    string `json:"dial_code"`
	PhoneNumber string `json:"phone_number"`
	Password    string `json:"password"`
}

type RegisterResponse struct {
	ID              uuid.UUID       `json:"id"`
	CreatedAt       time.Time       `json:"created_at"`
	UpdatedAt       time.Time       `json:"updated_at"`
	LastLoginAt     *time.Time      `json:"last_login_at"`
	PhoneNumber     string          `json:"phone_number"`
	CountryDialCode CountryDialCode `json:"country_dial_code"`
	Country         Country         `json:"country"`
	EasemobUsername string          `json:"easemob_username"`
	EasemobUUID     string          `json:"easemob_uuid"`
}

type SignInPayload struct {
	DialCode    string `json:"dial_code"`
	PhoneNumber string `json:"phone_number"`
	Password    string `json:"password"`
}

type RefreshTokenPayload struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}

type PairTokensResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type GetMe struct {
	ID              uuid.UUID       `json:"id"`
	CreatedAt       time.Time       `json:"created_at"`
	UpdatedAt       time.Time       `json:"updated_at"`
	DeletedAt       time.Time       `json:"deleted_at"`
	LastLoginAt     *time.Time      `json:"last_login_at"`
	PhoneNumber     string          `json:"phone_number"`
	CountryDialCode CountryDialCode `json:"country_dial_code"`
	Country         Country         `json:"country"`
	UserProfile     GetProfileRes   `json:"user_profile"`
	EasemobUsername string          `json:"easemob_username"`
	EasemobUUID     string          `json:"easemob_uuid"`
}